Added delayed auth timer for websocket connections
Delayed timer attempts to fix unintended side-effect of sleeping on the auth thread in situations where auth may have occurred too soon or has to be re-sent. Also increased websocket auth timeout from 10 seconds to 20 as a secondary measure for this as a delay in auth response does not necessarily mean a dead socket though too long of a timeout may negatively impact bot startup process with many shards.