Fixed thread count to be more resilient to deadlocks when not specified
Set 10 to be the default and set to 10 if 2 or fewer requested. Reasoning being that if a request is made that spawns a new async task, if the first waits on the second and the second spawns another task such as a REST request and waits on that, a deadlock will occur due to exhausting the thread pool. This can be a common scenario with a bot command that sends a channel message with a then continuation to send another message and waits for response while the command thread waits on the overall response. This will deadlock with only 2 threads. Ideally the count would be logical core count * 1 or * 2. (or more for systems with only 1-2 cores)