Added free(ish) functions for creating messages
Providing another param (channel_id) a message can be sent to any arbitrary channel assuming the bot is in it first. This can overcome any potential missing channel data as long as you know the id of the channel as well as make a potentially simpler method of sending a message to a channel.