Fixed potential race condition with deleting channels and guilds
Lib passes references around as much as it can and thus objects should persist for the entire lifetime of the lib's self.
While this can be considered a "memleak" in a sense because they are never destroyed for the lifetime of the lib,
for a small lib this is largely negligible.
Not fully ideal though. Something will need to be done later to guarantee the lifetime of objects while also clearing them out eventually.