Improve mutex acquisition for promise
Previously, the `l2` mutex was not guaranteed, as the `future` object could become invalid between check and lock. This would cause memory access violation.
Now the `l` and `l2` mutex is acquired from within a global mutex.

The work which happens within the `l` and `l2` locks now happens independently from the locking so that the global mutex is kept locked for minimal time.