Fix race condition during creation of `_self`
A race condition existed where the bot's `user` was created from `core::process_ready`, emplaced in `users`, and then another thread emplaced a user with the same ID. This was possible because in `core::process_ready` the common mutex guarding `users` was not acquired prior to emplacing our user.

Now the bot `user` is created as all users are, by using `user_create`. This will acquire the common mutex, and prevent such a race condition.